+++ bing/fitting/chisq_fit.py
