+++ bing/fitting/inference.py
