+++ papers/phytoplankton/Analysis/py/single_bing_fits.py
@@ -2,50 +2,13 @@
 import sys, os
 
 
+from bing import parameters as param20
+from bing.priors import standard 
+
 # Local
 sys.path.append(os.path.abspath("../../bing_2.0/Analysis/py"))
 import anly_utils_20
-import param as param20
 import dev_fits
-
-def standard_expb_pow():
-    # Priors
-    apriors=[dict(flavor='log_uniform', pmin=-6, pmax=5)]*3
-    bpriors=[dict(flavor='log_uniform', pmin=-6, pmax=5)]*2
-
-    # Uniform for Sdg from 0.01 - 0.02
-    apriors[1]=dict(flavor='uniform', pmin=0.01, pmax=0.02)
-
-    # Uniform for beta from 0. - 2. (positive here means negative slope)
-    bpriors[1]=dict(flavor='uniform', pmin=0., pmax=2.)
-
-    p = param20.p_ntuple(['ExpBricaud', 'Pow'], 
-        set_Sdg=False, sSdg=0.002, apriors=apriors, bpriors=bpriors,
-        scl_noise='PACE', nsteps=40000,
-        add_noise=True, wv_min=400., wv_max=700.)
-
-    return p
-
-def standard_giop():
-    # Priors
-    apriors=[dict(flavor='log_uniform', pmin=-6, pmax=5)]*2
-    bpriors=[dict(flavor='log_uniform', pmin=-6, pmax=5)]*1
-
-    p = param20.p_ntuple(['GIOP', 'Lee'], 
-        set_Sdg=False, sSdg=0.002, apriors=apriors, bpriors=bpriors,
-        scl_noise='PACE', nsteps=40000,
-        add_noise=True, wv_min=400., wv_max=700.)
-
-    return p
-
-def p_every_every(nsteps:int=500000):
-
-    p = param20.p_ntuple(['Every', 'Every'],
-        set_Sdg=False, sSdg=0.002, apriors=None, bpriors=None,
-        scl_noise=0.02, nsteps=nsteps,
-        add_noise=False, wv_min=400., wv_max=700.)
-
-    return p
 
 def main(flg):
     flg = int(flg)
@@ -54,14 +17,14 @@ def main(flg):
     if flg == 1:
 
         # Do it
-        p = standard_expb_pow()
+        p = standard.expb_pow()
         dev_fits.fit(p, 2773, show=True, seed=54321) 
 
     # High Chl + GIOP, Lee
     if flg == 2:
 
         # Do it
-        p = standard_giop()
+        p = standard.giop()
         dev_fits.fit(p, 2773, show=True, seed=54321)
 
     # High Chl + GSM
